refactor(graphql)!: `AstManipulationBuilderInfo` => `HandlerContextBuilderInfo`
