+++ packages/graphql/UPGRADE.md
@@ -91,5 +91,5 @@ This section is actual only if you are extending the package. Please review and 
 * [ ] To get `BuilderInfo` instance within Operator the `LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Context` should be used instead of `LastDragon_ru\LaraASP\GraphQL\Builder\Manipulator`:
 
     ```php
-    $context->get(\LastDragon_ru\LaraASP\GraphQL\Builder\Contexts\AstManipulationBuilderInfo::class)?->builderInfo
+    $context->get(LastDragon_ru\LaraASP\GraphQL\Builder\Directives\HandlerContextBuilderInfo::class)?->value
     ```
