+++ packages/graphql/src/Builder/Directives/HandlerContextBuilderInfo.php
@@ -1,10 +1,10 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\GraphQL\Builder\Contexts;
+namespace LastDragon_ru\LaraASP\GraphQL\Builder\Directives;
 
 use LastDragon_ru\LaraASP\GraphQL\Builder\BuilderInfo;
 
-class AstManipulationBuilderInfo {
+class HandlerContextBuilderInfo {
     public function __construct(
         public readonly BuilderInfo $value,
     ) {
