+++ packages/graphql/src/Builder/Directives/HandlerDirective.php
@@ -17,7 +17,6 @@ use Illuminate\Database\Query\Builder as QueryBuilder;
 use Laravel\Scout\Builder as ScoutBuilder;
 use LastDragon_ru\LaraASP\GraphQL\Builder\BuilderInfoDetector;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Context;
-use LastDragon_ru\LaraASP\GraphQL\Builder\Contexts\AstManipulationBuilderInfo;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Context as ContextContract;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Handler;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Operator;
@@ -45,7 +44,7 @@ use function is_array;
 use function reset;
 
 /**
- * @see AstManipulationBuilderInfo
+ * @see HandlerContextBuilderInfo
  */
 abstract class HandlerDirective extends BaseDirective implements Handler {
     use WithManipulator;
@@ -223,7 +222,7 @@ abstract class HandlerDirective extends BaseDirective implements Handler {
 
         // Argument
         $context = (new Context())->override([
-            AstManipulationBuilderInfo::class => new AstManipulationBuilderInfo($builder),
+            HandlerContextBuilderInfo::class => new HandlerContextBuilderInfo($builder),
         ]);
         $source  = $this->getFieldArgumentSource($manipulator, $parentType, $parentField, $argDefinition);
         $type    = $this->getArgDefinitionType($manipulator, $documentAST, $source, $context);
