+++ packages/graphql/src/Builder/Types/InputObject.php
@@ -9,12 +9,12 @@ use GraphQL\Language\AST\TypeDefinitionNode;
 use GraphQL\Language\BlockString;
 use GraphQL\Language\Parser;
 use GraphQL\Type\Definition\Type;
-use LastDragon_ru\LaraASP\GraphQL\Builder\Contexts\AstManipulationBuilderInfo;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Context;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Operator;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Scope;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\TypeDefinition;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\TypeSource;
+use LastDragon_ru\LaraASP\GraphQL\Builder\Directives\HandlerContextBuilderInfo;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Exceptions\TypeDefinitionFieldAlreadyDefined;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Manipulator;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Sources\InputFieldSource;
@@ -166,7 +166,7 @@ abstract class InputObject implements TypeDefinition {
         Context $context,
     ): ?InputValueDefinitionNode {
         // Builder?
-        $builder = $context->get(AstManipulationBuilderInfo::class)?->value->getBuilder();
+        $builder = $context->get(HandlerContextBuilderInfo::class)?->value->getBuilder();
 
         if (!$builder) {
             return null;
@@ -222,7 +222,7 @@ abstract class InputObject implements TypeDefinition {
         Context $context,
     ): ?Operator {
         // Builder?
-        $builder = $context->get(AstManipulationBuilderInfo::class)?->value->getBuilder();
+        $builder = $context->get(HandlerContextBuilderInfo::class)?->value->getBuilder();
 
         if (!$builder) {
             return null;
